+++ cli/src/commands/split.rs
@@ -24,6 +24,8 @@ use crate::cli_util::RevisionArg;
 use crate::command_error::user_error_with_hint;
 use crate::command_error::CommandError;
 use crate::complete;
+use crate::description_util::description_template;
+use crate::description_util::edit_description;
 use crate::description_util::edit_multiple_descriptions;
 use crate::ui::Ui;
 
@@ -66,6 +68,15 @@ pub(crate) struct SplitArgs {
     /// child
     #[arg(long, short)]
     parallel: bool,
+    /// Don't open an editor
+    ///
+    /// This keeps the description from the source commit as the description for
+    /// both new commits.
+    #[arg(long)]
+    no_edit: bool,
+    /// Only describe the first commit
+    #[arg(long)]
+    only_first: bool,
     /// Files matching any of these filesets are put in the first commit
     #[arg(
         value_name = "FILESETS",
@@ -139,7 +150,7 @@ The remainder will be in the second commit.
     let (first_commit, mut first_builder) = {
         let mut commit_builder = tx.repo_mut().rewrite_commit(&commit).detach();
         commit_builder.set_tree_id(selected_tree_id);
-        if commit_builder.description().is_empty() {
+        if commit_builder.description().is_empty() && !args.no_edit {
             commit_builder.set_description(tx.settings().get_string("ui.default-description")?);
         }
         (commit_builder.write_hidden()?, commit_builder)
@@ -171,29 +182,42 @@ The remainder will be in the second commit.
         (commit_builder.write_hidden()?, commit_builder)
     };
 
-    let commits = [
-        (
-            first_commit.id(),
-            &first_commit,
-            "Enter a description for the first commit.",
-        ),
-        (
-            second_commit.id(),
-            &second_commit,
-            "Enter a description for the second commit.",
-        ),
-    ];
-    let descriptions =
-        edit_multiple_descriptions(ui, &text_editor, &tx, commits)?.assert_complete()?;
+    if !args.no_edit {
+        if args.only_first {
+            let template = description_template(
+                ui,
+                &tx,
+                "Enter a description for the first commit.",
+                &first_commit,
+            )?;
+            let description = edit_description(&text_editor, &template)?;
+            first_builder.set_description(description);
+        } else {
+            let commits = [
+                (
+                    first_commit.id(),
+                    &first_commit,
+                    "Enter a description for the first commit.",
+                ),
+                (
+                    second_commit.id(),
+                    &second_commit,
+                    "Enter a description for the second commit.",
+                ),
+            ];
+            let descriptions =
+                edit_multiple_descriptions(ui, &text_editor, &tx, commits)?.assert_complete()?;
+            first_builder.set_description(&descriptions[first_commit.id()]);
+            second_builder.set_description(&descriptions[second_commit.id()]);
+        }
+    }
 
-    first_builder.set_description(&descriptions[first_commit.id()]);
     let first_commit = first_builder.write(tx.repo_mut())?;
 
     let second_commit = {
         if !args.parallel {
             second_builder.set_parents(vec![first_commit.id().clone()]);
         };
-        second_builder.set_description(&descriptions[second_commit.id()]);
         second_builder.write(tx.repo_mut())?
     };